requirements.txt
pandas
qrcode
Pillow